+++ changelog.md
@@ -1,5 +1,9 @@
 <h3>Changelog</h3>
 
+## 80 - Feb 20, 2016
+
+* Tempfix the player ui not resizing before being clicked.
+
 ## 79 - Feb 16, 2016
 
 * Tempfix the player ui not resizing before being clicked.